Consolidate interface projects so we can return some proper data types from IGame.
