+++ NexusMods.App.sln
@@ -8,8 +8,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.Paths.Tests", "te
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Game Locators", "Game Locators", "{954D1870-EE4B-4EC3-84AC-292C300461F0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.Interfaces", "src\NexusMods.Interfaces\NexusMods.Interfaces.csproj", "{0131FC55-53D2-4BE6-98DE-E55F7A133B1B}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.StandardGameLocators", "src\NexusMods.StandardGameLocators\NexusMods.StandardGameLocators.csproj", "{BB552259-24B0-4166-8A67-5649CE42BE83}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.StandardGameLocators.Tests", "tests\NexusMods.StandardGameLocators.Tests\NexusMods.StandardGameLocators.Tests.csproj", "{4FD95FB7-BD1D-4D79-90F8-9980DE5D6C28}"
@@ -17,8 +15,6 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Games", "Games", "{B95A2142-2872-4FFD-BAB2-596373401461}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.Games.BethesdaGameStudios", "src\Games\NexusMods.Games.BethesdaGameStudios\NexusMods.Games.BethesdaGameStudios.csproj", "{5038CF21-CE33-45F0-8C6F-D06D20D604F3}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.Games.Abstractions", "src\Games\Abstractions\NexusMods.Games.Abstractions\NexusMods.Games.Abstractions.csproj", "{0C8B3DC2-4DFD-455E-AA47-2CE5B9381BA9}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NexusMods.Games.BethesdaGameStudios.Tests", "tests\Games\NexusMods.Games.BethesdaGameStudios.Tests\NexusMods.Games.BethesdaGameStudios.Tests.csproj", "{DC3AE1FC-8E93-4F33-B937-C51C2F7C350F}"
 EndProject
@@ -80,10 +76,6 @@ Global
 		{1636CC83-8738-46ED-A688-D6AC2A55FEB5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1636CC83-8738-46ED-A688-D6AC2A55FEB5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1636CC83-8738-46ED-A688-D6AC2A55FEB5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0131FC55-53D2-4BE6-98DE-E55F7A133B1B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0131FC55-53D2-4BE6-98DE-E55F7A133B1B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0131FC55-53D2-4BE6-98DE-E55F7A133B1B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0131FC55-53D2-4BE6-98DE-E55F7A133B1B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BB552259-24B0-4166-8A67-5649CE42BE83}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BB552259-24B0-4166-8A67-5649CE42BE83}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BB552259-24B0-4166-8A67-5649CE42BE83}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -96,10 +88,6 @@ Global
 		{5038CF21-CE33-45F0-8C6F-D06D20D604F3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5038CF21-CE33-45F0-8C6F-D06D20D604F3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5038CF21-CE33-45F0-8C6F-D06D20D604F3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0C8B3DC2-4DFD-455E-AA47-2CE5B9381BA9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0C8B3DC2-4DFD-455E-AA47-2CE5B9381BA9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0C8B3DC2-4DFD-455E-AA47-2CE5B9381BA9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0C8B3DC2-4DFD-455E-AA47-2CE5B9381BA9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{DC3AE1FC-8E93-4F33-B937-C51C2F7C350F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DC3AE1FC-8E93-4F33-B937-C51C2F7C350F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DC3AE1FC-8E93-4F33-B937-C51C2F7C350F}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -187,7 +175,6 @@ Global
 		{BB552259-24B0-4166-8A67-5649CE42BE83} = {954D1870-EE4B-4EC3-84AC-292C300461F0}
 		{4FD95FB7-BD1D-4D79-90F8-9980DE5D6C28} = {954D1870-EE4B-4EC3-84AC-292C300461F0}
 		{5038CF21-CE33-45F0-8C6F-D06D20D604F3} = {B95A2142-2872-4FFD-BAB2-596373401461}
-		{0C8B3DC2-4DFD-455E-AA47-2CE5B9381BA9} = {B95A2142-2872-4FFD-BAB2-596373401461}
 		{DC3AE1FC-8E93-4F33-B937-C51C2F7C350F} = {B95A2142-2872-4FFD-BAB2-596373401461}
 		{3D21A286-CADC-40CF-AF29-790DAAD4363A} = {0D0CB8DD-34FE-42C0-9B53-4E27330F63D1}
 		{B4CA4927-6FFA-422B-8E54-413ECFCD6C39} = {0D0CB8DD-34FE-42C0-9B53-4E27330F63D1}
