+++ src/ArchiveManagement/NexusMods.FileExtractor/NexusMods.FileExtractor.csproj
@@ -10,7 +10,6 @@
       <ProjectReference Include="..\..\NexusMods.Common\NexusMods.Common.csproj" />
       <ProjectReference Include="..\..\NexusMods.DataModel.RateLimiting\NexusMods.DataModel.RateLimiting.csproj" />
       <ProjectReference Include="..\..\NexusMods.Hashing.xxHash64\NexusMods.Hashing.xxHash64.csproj" />
-      <ProjectReference Include="..\..\NexusMods.Interfaces\NexusMods.Interfaces.csproj" />
       <ProjectReference Include="..\..\NexusMods.Paths\NexusMods.Paths.csproj" />
     </ItemGroup>
