+++ src/Games/NexusMods.Games.BethesdaGameStudios/Installers/LooseFileInstaller.cs
@@ -1,5 +1,6 @@
 ﻿using NexusMods.DataModel.Abstractions;
 using NexusMods.DataModel.ArchiveContents;
+using NexusMods.DataModel.Games;
 using NexusMods.DataModel.ModInstallers;
 using NexusMods.DataModel.Loadouts;
 using NexusMods.DataModel.Loadouts.ModFiles;
