+++ src/Games/NexusMods.Games.BethesdaGameStudios/Services.cs
@@ -1,10 +1,10 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using NexusMods.Common;
 using NexusMods.DataModel.Abstractions;
+using NexusMods.DataModel.Games;
 using NexusMods.DataModel.JsonConverters.ExpressionGenerator;
 using NexusMods.DataModel.ModInstallers;
 using NexusMods.Games.BethesdaGameStudios.Installers;
-using NexusMods.Interfaces;
-using NexusMods.Interfaces.Components;
 
 namespace NexusMods.Games.BethesdaGameStudios;
