+++ src/Games/NexusMods.Games.BethesdaGameStudios/SkyrimSpecialEdition.cs
@@ -1,8 +1,7 @@
 ﻿using Microsoft.Extensions.Logging;
+using NexusMods.DataModel.Games;
 using NexusMods.Games.Abstractions;
 using NexusMods.Interfaces;
-using NexusMods.Interfaces.Components;
-using NexusMods.Interfaces.StoreLocatorTags;
 using NexusMods.Paths;
 using NexusMods.StandardGameLocators;
