+++ src/Games/NexusMods.Games.DarkestDungeon/DarkestDungeonModInstaller.cs
@@ -1,5 +1,6 @@
 ﻿using NexusMods.DataModel.Abstractions;
 using NexusMods.DataModel.ArchiveContents;
+using NexusMods.DataModel.Games;
 using NexusMods.DataModel.ModInstallers;
 using NexusMods.DataModel.Loadouts;
 using NexusMods.DataModel.Loadouts.ModFiles;
