+++ src/NexusMods.Common/Priority.cs
