+++ src/NexusMods.DataModel/ArchiveManager.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.Extensions.Logging;
+using NexusMods.Common;
 using NexusMods.DataModel.Abstractions;
 using NexusMods.DataModel.ArchiveContents;
 using NexusMods.DataModel.RateLimiting;
 using NexusMods.Hashing.xxHash64;
-using NexusMods.Interfaces.Streams;
 using NexusMods.Paths;
 
 namespace NexusMods.DataModel;
