+++ src/NexusMods.DataModel/Games/IGameLocator.cs
@@ -1,6 +1,4 @@
-﻿using NexusMods.Paths;
-
-namespace NexusMods.Interfaces.Components;
+﻿namespace NexusMods.DataModel.Games;
 
 /// <summary>
 /// A service capable of finding the path to the game's installation directory.
