+++ src/NexusMods.DataModel/Games/ISteamGame.cs
@@ -1,6 +1,4 @@
-﻿using NexusMods.Interfaces.Components;
-
-namespace NexusMods.Interfaces.StoreLocatorTags;
+﻿namespace NexusMods.DataModel.Games;
 
 /// <summary>
 /// Marker interface for the steam store locator.
