+++ src/NexusMods.DataModel/Loadouts/Loadout.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections;
 using NexusMods.DataModel.Abstractions;
+using NexusMods.DataModel.Games;
 using NexusMods.DataModel.JsonConverters;
 using NexusMods.Interfaces;
 using NexusMods.Paths;
