+++ src/NexusMods.DataModel/Loadouts/ModFiles/GameFile.cs
@@ -1,4 +1,5 @@
-﻿using NexusMods.DataModel.JsonConverters;
+﻿using NexusMods.DataModel.Games;
+using NexusMods.DataModel.JsonConverters;
 using NexusMods.Hashing.xxHash64;
 using NexusMods.Interfaces;
 using NexusMods.Paths;
