+++ src/NexusMods.DataModel/ModInstallers/IModInstaller.cs
@@ -1,5 +1,6 @@
 ﻿using NexusMods.DataModel.Abstractions;
 using NexusMods.DataModel.ArchiveContents;
+using NexusMods.DataModel.Games;
 using NexusMods.DataModel.Loadouts;
 using NexusMods.Hashing.xxHash64;
 using NexusMods.Interfaces;
