+++ src/NexusMods.StandardGameLocators/AGameLocator.cs
@@ -1,7 +1,6 @@
 ﻿using GameFinder.Common;
 using Microsoft.Extensions.Logging;
-using NexusMods.Interfaces;
-using NexusMods.Interfaces.Components;
+using NexusMods.DataModel.Games;
 using NexusMods.Paths;
 
 namespace NexusMods.StandardGameLocators;
