+++ src/NexusMods.StandardGameLocators/GogLocator.cs
@@ -1,7 +1,7 @@
 ﻿using GameFinder.Common;
 using GameFinder.StoreHandlers.GOG;
 using Microsoft.Extensions.Logging;
-using NexusMods.Interfaces.StoreLocatorTags;
+using NexusMods.DataModel.Games;
 using NexusMods.Paths;
 
 namespace NexusMods.StandardGameLocators;
