+++ tests/NexusMods.StandardGameLocators.Tests/NexusMods.StandardGameLocators.Tests.csproj
@@ -30,7 +30,6 @@
 
     <ItemGroup>
       <ProjectReference Include="..\..\src\NexusMods.DataModel\NexusMods.DataModel.csproj" />
-      <ProjectReference Include="..\..\src\NexusMods.Interfaces\NexusMods.Interfaces.csproj" />
       <ProjectReference Include="..\..\src\NexusMods.Paths\NexusMods.Paths.csproj" />
       <ProjectReference Include="..\..\src\NexusMods.StandardGameLocators\NexusMods.StandardGameLocators.csproj" />
       <ProjectReference Include="..\..\src\NexusMods.StandardGameLocators\NexusMods.StandardGameLocators.csproj" />
